--- data/src/main/java/com/team_ia/data/remote/datasource/AdminDataSource.kt
@@ -1,4 +1,0 @@
-package com.team_ia.data.remote.datasource
-
-interface AdminDataSource {
-}
