--- data/src/main/java/com/team_ia/data/remote/datasource/PostDataSource.kt
@@ -1,4 +1,0 @@
-package com.team_ia.data.remote.datasource
-
-interface PostDataSource {
-}
