+++ data/src/main/java/com/team_ia/data/remote/datasource/application/ApplicationDataSourceImpl.kt
@@ -1,4 +1,4 @@
-package com.team_ia.data.remote.datasource
+package com.team_ia.data.remote.datasource.application
 
 import com.team_ia.data.remote.api.ApplicationAPI
 import javax.inject.Inject
