+++ data/src/main/java/com/team_ia/data/remote/datasource/auth/AuthDataSource.kt
@@ -1,0 +1,4 @@
+package com.team_ia.data.remote.datasource.auth
+
+interface AuthDataSource {
+}
