+++ data/src/main/java/com/team_ia/data/remote/datasource/auth/AuthDataSourceImpl.kt
@@ -1,4 +1,4 @@
-package com.team_ia.data.remote.datasource
+package com.team_ia.data.remote.datasource.auth
 
 import com.team_ia.data.remote.api.AuthAPI
 import javax.inject.Inject
