+++ data/src/main/java/com/team_ia/data/remote/datasource/img/ImgDataSourceImpl.kt
@@ -1,4 +1,4 @@
-package com.team_ia.data.remote.datasource
+package com.team_ia.data.remote.datasource.img
 
 import com.team_ia.data.remote.api.ImgAPI
 import javax.inject.Inject
