+++ data/src/main/java/com/team_ia/data/remote/datasource/member/MemberDataSourceImpl.kt
@@ -1,4 +1,4 @@
-package com.team_ia.data.remote.datasource
+package com.team_ia.data.remote.datasource.member
 
 import com.team_ia.data.remote.api.MemberAPI
 import javax.inject.Inject
