+++ data/src/main/java/com/team_ia/data/remote/datasource/post/PostDataSource.kt
@@ -1,0 +1,4 @@
+package com.team_ia.data.remote.datasource.post
+
+interface PostDataSource {
+}
