+++ data/src/main/java/com/team_ia/data/remote/datasource/post/PostDataSourceImpl.kt
@@ -1,4 +1,4 @@
-package com.team_ia.data.remote.datasource
+package com.team_ia.data.remote.datasource.post
 
 import com.team_ia.data.remote.api.PostAPI
 import javax.inject.Inject
