+++ data/src/main/java/com/team_ia/data/repository/AdminRepositoryImpl.kt
@@ -1,6 +1,6 @@
 package com.team_ia.data.repository
 
-import com.team_ia.data.remote.datasource.AdminDataSource
+import com.team_ia.data.remote.datasource.admin.AdminDataSource
 import com.team_ia.domain.repository.AdminRepository
 import javax.inject.Inject
