+++ data/src/main/java/com/team_ia/data/repository/ApplicationRepositoryImpl.kt
@@ -1,6 +1,6 @@
 package com.team_ia.data.repository
 
-import com.team_ia.data.remote.datasource.ApplicationDataSource
+import com.team_ia.data.remote.datasource.application.ApplicationDataSource
 import com.team_ia.domain.repository.ApplicationRepository
 import javax.inject.Inject
