+++ data/src/main/java/com/team_ia/data/repository/AuthRepositoryImpl.kt
@@ -1,6 +1,6 @@
 package com.team_ia.data.repository
 
-import com.team_ia.data.remote.datasource.AuthDataSource
+import com.team_ia.data.remote.datasource.auth.AuthDataSource
 import com.team_ia.domain.repository.AuthRepository
 import javax.inject.Inject
