+++ data/src/main/java/com/team_ia/data/repository/EmailRepositoryImpl.kt
@@ -1,6 +1,6 @@
 package com.team_ia.data.repository
 
-import com.team_ia.data.remote.datasource.EmailDataSource
+import com.team_ia.data.remote.datasource.email.EmailDataSource
 import com.team_ia.domain.repository.EmailRepository
 import javax.inject.Inject
