+++ data/src/main/java/com/team_ia/data/repository/ImgRepositoryImpl.kt
@@ -1,6 +1,6 @@
 package com.team_ia.data.repository
 
-import com.team_ia.data.remote.datasource.ImgDataSource
+import com.team_ia.data.remote.datasource.img.ImgDataSource
 import com.team_ia.domain.repository.ImgRepository
 import javax.inject.Inject
