+++ data/src/main/java/com/team_ia/data/repository/MemberRepositoryImpl.kt
@@ -1,6 +1,6 @@
 package com.team_ia.data.repository
 
-import com.team_ia.data.remote.datasource.MemberDataSource
+import com.team_ia.data.remote.datasource.member.MemberDataSource
 import com.team_ia.domain.repository.MemberRepository
 import javax.inject.Inject
