+++ data/src/main/java/com/team_ia/data/repository/PostRepositoryImpl.kt
@@ -1,6 +1,6 @@
 package com.team_ia.data.repository
 
-import com.team_ia.data.remote.datasource.PostDataSource
+import com.team_ia.data.remote.datasource.post.PostDataSource
 import com.team_ia.domain.repository.PostRepository
 import javax.inject.Inject
