+++ data/src/main/java/com/team_ia/data/repository/SocialRepositoryImpl.kt
@@ -1,6 +1,6 @@
 package com.team_ia.data.repository
 
-import com.team_ia.data.remote.datasource.SocialDataSource
+import com.team_ia.data.remote.datasource.social.SocialDataSource
 import com.team_ia.domain.repository.SocialRepository
 import javax.inject.Inject
